refactor: remove duplicate factory logic from Validator facade
Validator facade now delegates all creation to ValidatorFactory
instead of maintaining its own VALIDATOR_MAP copy. This eliminates
the DRY violation — only one place to update when adding validators.

Also updated both files to use Domain\Contracts namespace for
ValidatorInterface and ValidatorFactoryInterface.

Converted ValidatorFactory from switch/case to VALIDATOR_MAP
constant for consistency and maintainability.
