+++ src/Application/Validators/ValidatorFactory.php
@@ -2,6 +2,8 @@
 
 namespace QDenka\EasyValidation\Application\Validators;
 
+use QDenka\EasyValidation\Domain\Contracts\ValidatorFactoryInterface;
+use QDenka\EasyValidation\Domain\Contracts\ValidatorInterface;
 use QDenka\EasyValidation\Domain\Ip\IpValidator;
 use QDenka\EasyValidation\Domain\Uuid\UuidValidator;
 use QDenka\EasyValidation\Domain\Json\JsonValidator;
@@ -15,37 +17,38 @@ use QDenka\EasyValidation\Domain\Number\NumberValidator;
 use QDenka\EasyValidation\Domain\Url\UrlValidator;
 use QDenka\EasyValidation\Infrastructure\Email\FileDisposableEmailDomainProvider;
 
-
 class ValidatorFactory implements ValidatorFactoryInterface
 {
+    private const VALIDATOR_MAP = [
+        'email'             => EmailValidator::class,
+        'google_email'      => GoogleMailValidator::class,
+        'disposable_email'  => DisposableEmailValidator::class,
+        'url'               => UrlValidator::class,
+        'number'            => NumberValidator::class,
+        'date'              => DateValidator::class,
+        'ip'                => IpValidator::class,
+        'uuid'              => UuidValidator::class,
+        'json'              => JsonValidator::class,
+        'base64'            => Base64Validator::class,
+        'phone'             => PhoneNumberValidator::class,
+    ];
+
     public static function create(string $type): ?ValidatorInterface
     {
-        switch (strtolower($type)) {
-            case 'email':
-                return new EmailValidator();
-            case 'google_email':
-                return new GoogleMailValidator();
-            case 'disposable_email':
-                $provider = new FileDisposableEmailDomainProvider(__DIR__ . '/../../../config/disposable_domains.php');
-                return new DisposableEmailValidator($provider);
-            case 'date':
-                return new DateValidator();
-            case 'number':
-                return new NumberValidator();
-            case 'url':
-                return new UrlValidator();
-            case 'ip':
-                return new IpValidator();
-            case 'uuid':
-                return new UuidValidator();
-            case 'json':
-                return new JsonValidator();
-            case 'base64':
-                return new Base64Validator();
-            case 'phone':
-                return new PhoneNumberValidator();
-            default:
-                return null;
+        $type = strtolower($type);
+
+        if (!array_key_exists($type, self::VALIDATOR_MAP)) {
+            return null;
         }
+
+        $class = self::VALIDATOR_MAP[$type];
+
+        if ($type === 'disposable_email') {
+            $configPath = __DIR__ . '/../../../config/disposable_domains.php';
+            $provider = new FileDisposableEmailDomainProvider($configPath);
+            return new $class($provider);
+        }
+
+        return new $class();
     }
 }
